Move IncCompilationContext to codegen package, rename it to IncrementalCodeGenContext
